+++ player/playloop.c
@@ -550,7 +550,7 @@ double get_current_pos_ratio(struct MPContext *mpctx, bool use_range)
 int get_percent_pos(struct MPContext *mpctx)
 {
     double pos = get_current_pos_ratio(mpctx, false);
-    return pos < 0 ? -1 : pos * 100;
+    return pos < 0 ? -1 : (int) (pos * 100);
 }
 
 // -2 is no chapters, -1 is before first chapter
@@ -683,7 +683,7 @@ static void handle_pause_on_low_cache(struct MPContext *mpctx)
 
     if (mpctx->paused_for_cache) {
         cache_buffer =
-            100 * MPCLAMP(s.ts_duration / opts->cache_pause_wait, 0, 0.99);
+            (int) (100 * MPCLAMP(s.ts_duration / opts->cache_pause_wait, 0, 0.99));
         mp_set_timeout(mpctx, 0.2);
     }
